started Finite State Machine
https://www.youtube.com/watch?v=5PTd0WdKB-4
following that tutorial to make a play system

The idea behind this is that each robot will move forward and then raise the endmove event.  The state will be in charge of sending the start move and determine behavior after moving.

This is mostly just setting up the skeleton though so this isn't implimented yet